+++ just_audio_background/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.1-beta.8
+
+* Implement setPitch, dispose, setSkipSilence.
+* Implement audio effect methods.
+* Implement setPreferredPeakBitRate.
+* Implement setCanUseNetworkResourcesForLiveStreamingWhilePaused.
+
 ## 0.0.1-beta.7
 
 * Fix build when targeting Android 13.